+++ contracts/interfaces/IdLogos.sol
@@ -1,11 +1,13 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.24;
 
+import {SplitV2Lib} from "../splitsV2/libraries/SplitV2.sol";
+
 interface IDLogos {
     struct Backer {
         address addr;
+        address referrer;
         uint256 amount;
-        bool votesToReject;
     }
 
     enum SpeakerStatus {
@@ -16,7 +18,7 @@ interface IDLogos {
 
     struct Speaker {
         address addr;
-        uint16 fee; // in BPS
+        uint256 fee;
         string provider; // e.g. X, Discord etc.
         string handle;
         SpeakerStatus status;
@@ -37,72 +39,102 @@ interface IDLogos {
         string title;
         string mediaAssetURL;
         address proposer;
-        uint scheduledAt;
+        uint256 proposerFee;
+        uint256 scheduledAt;
         uint256 minimumPledge;
-        uint crowdfundStartAt;
-        uint crowdfundEndAt;
-        uint rejectionDeadline;
-        address splits; // Splits Address
+        uint256 crowdfundStartAt;
+        uint256 crowdfundEndAt;
+        uint256 rejectionDeadline;
+        address splitForSpeaker; // PUshSplit address for dlogos, community, speakers
+        address splitForAffiliate; // PUshSplit address for affiliates
         Status status;
     }
 
     /// EVENTS
+    event DLogosAddressUpdated(address _dLogos);
+    event CommunityAddressUpdated(address _community);
+    event DLogosFeeUpdated(uint256 _dLogosFee);
+    event CommunityFeeUpdated(uint256 _communityFee);
+    event AffiliateFeeUpdated(uint256 _affiliateFee);
     event RejectThresholdUpdated(uint16 indexed _fee);
-    event DurationThresholdUpdated(uint8 _durationThreshold);
+    event MaxDurationUpdated(uint8 _maxDuration);
+    event RejectinoWindowUpdated(uint8 _rejectionWindow);
+    event ZeroFeeProposersSet(address[] _proposers, bool[] _statuses);
     event LogoCreated(
         address indexed _owner,
-        uint indexed _logoId,
-        uint indexed _crowdfundStartAt
+        uint256 indexed _logoId,
+        uint256 indexed _crowdfundStartAt
     );
-    event MinimumPledgeSet(address indexed _owner, uint indexed _minimumPledge);
-    event Crowdfund(address indexed _owner, uint indexed _amount);
+    event ProposerFeeUpdated(address indexed _proposer, uint256 indexed _logoId, uint256 _proposerFee);
+    event MinimumPledgeSet(address indexed _owner, uint256 indexed _minimumPledge);
+    event Crowdfund(
+        uint256 indexed _logoId,
+        address indexed _owner, 
+        uint256 indexed _amount
+    );
     event CrowdfundToggled(
         address indexed _owner,
         bool indexed _crowdfundIsOpen
     );
-    event FundsWithdrawn(address indexed _owner, uint indexed _amount);
+    event FundsWithdrawn(address indexed _owner, uint256 indexed _amount);
     event SpeakersSet(
         address indexed _owner,
         address[] _speakers,
-        uint16[] _fees,
+        uint256[] _fees,
         string[] _providers,
         string[] _handles
     );
-    event DateSet(address indexed _owner, uint indexed _scheduledAt);
+    event DateSet(address indexed _owner, uint256 indexed _scheduledAt);
     event MediaAssetSet(address indexed _owner, string indexed _mediaAssetURL);
-    event SplitsSet(address indexed _owner, address indexed _splitsAddress);
     event RewardsDistributed(
-        address indexed _owner,
-        address indexed _splitsAddress,
-        uint256 indexed _totalRewards
+        address indexed _proposer, 
+        address _splitForSpeaker, 
+        address _splitForAffiliate, 
+        uint256 _totalRewards
     );
-    event SpeakerStatusSet(uint indexed _logoId, address indexed _speaker, uint indexed _status);
-    event RejectionSubmitted(uint indexed _logoId, address indexed _backer);
+    event SpeakerStatusSet(uint256 indexed _logoId, address indexed _speaker, uint256 indexed _status);
+    event RejectionSubmitted(uint256 indexed _logoId, address indexed _backer);
     event RefundInitiated(
-        uint indexed _logoId, 
+        uint256 indexed _logoId, 
         bool _case1,
         bool _case2,
         bool _case3,
         bool _case4
+    );
+    event SplitForAffiliateCreated(
+        address indexed _split, 
+        SplitV2Lib.Split _splitParams, 
+        address _owner, 
+        address _creator
+    );
+    event SplitForSpeakerCreated(
+        address indexed _split, 
+        SplitV2Lib.Split _splitParams, 
+        address _owner, 
+        address _creator
     );
 
     function logoId() external view returns (uint256);
 
     function rejectThreshold() external view returns (uint16);
 
-    function durationThreshold() external view returns (uint8);
+    function maxDuration() external view returns (uint8);
 
-    function initialize() external;  
+    function initialize(
+        address _pushSplitFactory,
+        address _dLogos,
+        address _community
+    ) external;  
 
     function setRejectThreshold(uint16) external;
 
-    function setDurationThreshold(uint8) external;
+    function setMaxDuration(uint8) external;
 
-    function createLogo(string calldata, uint8) external returns (uint256);  
+    function createLogo(uint256, string calldata, uint8) external returns (uint256);  
 
     function toggleCrowdfund(uint256) external;
 
-    function crowdfund(uint256) external payable;
+    function crowdfund(uint256, address) external payable;
 
     function setMinimumPledge(uint256, uint256) external;
 
@@ -117,14 +149,14 @@ interface IDLogos {
     function setSpeakers(
         uint256, 
         address[] calldata, 
-        uint16[] calldata, 
+        uint256[] calldata, 
         string[] calldata, 
         string[] calldata
     ) external;
 
     function setSpeakerStatus(
         uint256,
-        uint256
+        uint8
     ) external;
 
     function getSpeakersForLogo(uint256) external view returns (Speaker[] memory);
@@ -140,8 +172,7 @@ interface IDLogos {
     ) external;
 
     function distributeRewards(
-        uint256,
-        address
+        uint256
     ) external;
 
     function pause() external;